Create README.md
The pet industry has been and is still one of the most lucrative of all. The number of people owning pets increases every year and so are the feeding demands of the pets. This project aims at solving the problems faced by pet owners by allowing them to customize the amount of food delivered and the time interval between feeding times. At the end of each month the summary of the pet's feeding is sent to a database thereby allowing the owner to plan efficiently.